% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setObj.R
\name{addAtacObj}
\alias{addAtacObj}
\title{Creat additional slots for atac-seq analysis extended from an existing object}
\usage{
addAtacObj(
  obj,
  pmat = NULL,
  fragments = NULL,
  promoterRegion = NULL,
  TSS = NULL,
  blacklist = NULL,
  exons = NULL,
  transcripts = NULL,
  enhancers = NULL,
  tool = "pagoda2"
)
}
\arguments{
\item{obj}{existing object (paagoda2)}

\item{pmat}{peak matrix - rows are cells colums are peaks (chr:start-end)}

\item{promoterRegion}{GRanges object containing the promoter regions}

\item{TSS}{GRanges object containing tss locations}

\item{frament}{fragment file that has the same format as
https://support.10xgenomics.com/single-cell-atac/software/pipelines/latest/output/fragments}
}
\description{
Creat additional slots for atac-seq analysis extended from an existing object
}
