% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setObj.R
\name{addMotifMatrix}
\alias{addMotifMatrix}
\title{generate motif to feature(peak) matrix, row motifs, column matching score for a feature(peak)}
\usage{
addMotifMatrix(
  obj,
  genome = "BSgenome.Hsapiens.NCBI.GRCh38",
  outype = "matches"
)
}
\arguments{
\item{obj}{pagoda2 object}

\item{outype}{types to count motifs (scores, matches)}
}
\value{
obj with motif block
}
\description{
generate motif to feature(peak) matrix, row motifs, column matching score for a feature(peak)
}
