% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{assignPeakToGene}
\alias{assignPeakToGene}
\title{asssign peak to its nearest gene}
\usage{
assignPeakToGene(
  obj,
  use.pvalues = FALSE,
  cutoff = 0.05,
  test.asso = TRUE,
  n.cores = 10
)
}
\arguments{
\item{cutoff, }{adjPvalue (Pvalue) cutoff}

\item{test.asso}{TRUE perform statistical test the significance of identified peak-gene pairs}

\item{use.pvalue}{TRUE, use pvalue to identify DAR region - default FALSE}
}
\description{
asssign peak to its nearest gene
}
