% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet.R
\name{calDoubletScore}
\alias{calDoubletScore}
\title{calculate score for potential doublet}
\usage{
calDoubletScore(
  obj,
  plot = T,
  plotDoubleEmbedding = T,
  plotDepthDistr = T,
  k = 30
)
}
\description{
calculate score for potential doublet
}
