% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calJaccard}
\alias{calJaccard}
\title{calculate jaccard index between two binary matrices}
\usage{
calJaccard(m1, m2)
}
\arguments{
\item{m1}{matrix 1}

\item{m2}{matrix 2}
}
\description{
calculate jaccard index between two binary matrices
}
