% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callMACs}
\alias{callMACs}
\title{call peaks based on macs2
obj: p2 object record the fragment file paths}
\usage{
callMACs(
  fragmentFileDir,
  output.prefix = "",
  path.to.macs,
  gsize,
  buffer.size = 500,
  out.dir = getwd(),
  macs.options = "--nomodel --shift 37 --ext 73 --qval 1e-2 -B --SPMR --call-summits",
  keep.simple = TRUE
)
}
\arguments{
\item{macs.options:}{macs option for peak calling - adapted from snapATAC option
(https://github.com/r3fang/SnapATAC/blob/master)}
}
\description{
call peaks based on macs2
obj: p2 object record the fragment file paths
}
