% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callMACsCluster}
\alias{callMACsCluster}
\title{call peaks based on assigned cell clusters}
\usage{
callMACsCluster(
  obj,
  clusters = NULL,
  path.to.macs,
  gsize,
  buffer.size = 500,
  out.dir = getwd(),
  macs.options = "--nomodel --shift 37 --ext 73 --qval 1e-2 -B --SPMR --call-summits",
  ncores = 1,
  keep.simple = TRUE,
  min.cells = 10,
  includeOthers = TRUE
)
}
\arguments{
\item{obj}{p2 object}

\item{clusters}{cell annotation}
}
\description{
call peaks based on assigned cell clusters
}
