% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{cellStatistics}
\alias{cellStatistics}
\title{calculate QC statistics per each cell
 blacklist ratio, promoter ratio, fragment stat}
\usage{
cellStatistics(obj, overallStat = TRUE, n.cores = 20)
}
\description{
calculate QC statistics per each cell
 blacklist ratio, promoter ratio, fragment stat
}
