% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{ciceroGeneActivity}
\alias{ciceroGeneActivity}
\title{calculate cicero gene activity score}
\usage{
ciceroGeneActivity(input_cds, genome = "hg38", genomeSize)
}
\arguments{
\item{input_cds}{processed cicero object}

\item{geneomeSize}{data.frame contains chromosome sizes}
}
\description{
calculate cicero gene activity score
}
