% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{createInputCDS}
\alias{createInputCDS}
\title{creating input CDS from custom peaks}
\usage{
createInputCDS(p2 = NULL, pmat = NULL)
}
\arguments{
\item{p2}{pagoda 2 object}

\item{pmat}{cell by peak matrix}
}
\description{
creating input CDS from custom peaks
}
