% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dimReduct}
\alias{dimReduct}
\title{run dimension reduction}
\usage{
dimReduct(
  obj,
  matrix = "npmat",
  method = "svd",
  dm = 50,
  scale.embedding = TRUE
)
}
\arguments{
\item{obj}{p2 object}

\item{matrix}{matrix for dimension reduction (pmat, npmat, jmat, njmat) for ATAC data}
}
\description{
run dimension reduction
}
