% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{enrichMotifs}
\alias{enrichMotifs}
\title{identify enriched motifs for a feature set (peaks) based on query regions}
\usage{
enrichMotifs(obj, targetRegion, sampleMethod = "match.GC", nsample = 10000)
}
\arguments{
\item{targetRegion}{genomic region sets for do enrichment analysis (e.g. chr17:44069893-44071328)}

\item{queryRegion}{genomic regions for all features (peaks)}
}
\description{
identify enriched motifs for a feature set (peaks) based on query regions
}
