% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{fragmentOverlapPeaks}
\alias{fragmentOverlapPeaks}
\title{Caculate the number of fragment that overlapped with peak region and return summary statistics
  (median_total_usable_fragments_per_cell, median_unique_fragments_per_cell)}
\usage{
fragmentOverlapPeaks(obj, plot = T, fitModel = FALSE, n.cores = 10)
}
\arguments{
\item{obj}{an class object (default pagoda2)}

\item{fitModel}{Whether to estimate fragments assuming fragment contamination rate is 0.02}
}
\value{
plot and summary statistics
}
\description{
Caculate the number of fragment that overlapped with peak region and return summary statistics
  (median_total_usable_fragments_per_cell, median_unique_fragments_per_cell)
}
