% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{GetMarkerExprMatrix}
\alias{GetMarkerExprMatrix}
\title{Get gene expression matrix of associated marker genes}
\usage{
GetMarkerExprMatrix(count, groups, genes, scale = TRUE, logscale = TRUE)
}
\arguments{
\item{count}{raw count matrix - rows are cells, columns are genes}

\item{groups}{cluster annotation}

\item{genes}{genelist}

\item{scale}{TRUE scale the data}
}
\description{
Get gene expression matrix of associated marker genes
}
