% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{hvgPeaks}
\alias{hvgPeaks}
\title{calculate high variable peaks based on normalized cell by peak matrix}
\usage{
hvgPeaks(obj, percentileCutoff = NULL)
}
\arguments{
\item{obj}{pagoda2 object}

\item{percentileCutoff}{cutoff value for high variable peaks (0-1)}
}
\description{
calculate high variable peaks based on normalized cell by peak matrix
}
