% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{idenDAR}
\alias{idenDAR}
\title{identify differential accessible region based on pagoda object}
\usage{
idenDAR(
  obj,
  clusters,
  neg = NULL,
  method = c("knn", "random"),
  bcv = 0.1,
  test.method = c("exactTest", "lrt", "qlf"),
  n.cores = 2,
  bgCellnumber = NULL
)
}
\arguments{
\item{obj}{pagoda object}

\item{clusters}{factor contain cluster information for each cell}

\item{bgCellnumer}{background cell number}
}
\description{
identify differential accessible region based on pagoda object
}
