% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{intPosMat}
\alias{intPosMat}
\title{generate integrated cell position matrix given sets of regions}
\usage{
intPosMat(
  obj,
  regions,
  cells = NULL,
  sampleRegion = T,
  nsamples = 5000,
  n.cores = 10
)
}
\arguments{
\item{regions}{regions(GRanges object) that used to generate cell position matrix}
}
\value{
position matrix
}
\description{
generate integrated cell position matrix given sets of regions
}
