% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normJD}
\alias{normJD}
\title{normalize jaccard index matrix based on fang et al
  doi: https://doi.org/10.1101/615179}
\usage{
normJD(obj, method = "residual", center = TRUE)
}
\arguments{
\item{obj}{default: pagoda object}
}
\value{
obj with normalized jaccard matrix
}
\description{
normalize jaccard index matrix based on fang et al
  doi: https://doi.org/10.1101/615179
}
