% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normPmatrix}
\alias{normPmatrix}
\title{normalize peak matrix based on TF-IDF and other algorithms}
\usage{
normPmatrix(obj, method = "TfIDF", scale.factor = 10000)
}
\arguments{
\item{obj}{p2 object}

\item{method}{TfIDF: LSI TF-IDF normalization in Stuart & Butler et al. 2019
log-TfIDF: The log-TF method
TF: IDF normalization no TF}
}
\value{
normalized peak matrix
}
\description{
normalize peak matrix based on TF-IDF and other algorithms
}
