% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{plotAssoGenes}
\alias{plotAssoGenes}
\title{plot heatmap of associated genes}
\usage{
plotAssoGenes(
  obj,
  pal = colorRampPalette(c("dodgerblue1", "grey95", "indianred1"))(1024),
  distanceCutoff = -1000,
  use.pvalue.cutoff = 0.05,
  RNAcluster = NULL
)
}
\description{
plot heatmap of associated genes
}
