% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{plotComponentVariancePagoda}
\alias{plotComponentVariancePagoda}
\title{plotComponentVariance based on pagoda obj}
\usage{
plotComponentVariancePagoda(obj, space = "PCA", dm = 20)
}
\arguments{
\item{obj}{pagoda object}

\item{dm}{dimensions to plot}
}
\description{
plotComponentVariance based on pagoda obj
}
