% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotDARoverlap}
\alias{plotDARoverlap}
\title{plot distribution of DAR overlapping with annotated regions}
\usage{
plotDARoverlap(obj, use.pvalues = FALSE, cutoff = 0.05)
}
\arguments{
\item{obj}{p2 object}

\item{use.pvalues}{use pvalue to identify DAR region}

\item{cutoff}{p value (adj pvalue) cutoff}
}
\description{
plot distribution of DAR overlapping with annotated regions
}
