% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{plotFeature}
\alias{plotFeature}
\title{plot features based on embedding for individual cluster}
\usage{
plotFeature(
  obj,
  clusterID = NULL,
  features = "DAR",
  genename = NULL,
  main = NULL,
  PvalCutoff = FALSE,
  cutoff = 0.05
)
}
\arguments{
\item{obj}{pagoda or related object}
}
\description{
plot features based on embedding for individual cluster
}
