% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotFragmentScatter}
\alias{plotFragmentScatter}
\title{plotting fragment distribution
 Scatterplot showing the number of fragments/barcode and the percent of fragments overlapping peaks}
\usage{
plotFragmentScatter(obj)
}
\description{
plotting fragment distribution
 Scatterplot showing the number of fragments/barcode and the percent of fragments overlapping peaks
}
