% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotSnareBarcodeComparison}
\alias{plotSnareBarcodeComparison}
\title{plot the comparison between RNA and ATAC barcode}
\usage{
plotSnareBarcodeComparison(obj)
}
\arguments{
\item{obj}{pagoda2 object}
}
\description{
plot the comparison between RNA and ATAC barcode
}
