% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{posMatrix}
\alias{posMatrix}
\title{Generate cell position matrix in a given region based on middle position of a fragment}
\usage{
posMatrix(obj, region, cells = NULL, fragments = NULL)
}
\arguments{
\item{obj}{an R object (default:pagoda2)}

\item{region}{regions of interest (GRanges object)}

\item{cells}{which cell to include. Default: all cells in the object}
}
\description{
Generate cell position matrix in a given region based on middle position of a fragment
}
