% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{PrepareFragmentPlot}
\alias{PrepareFragmentPlot}
\title{prepare data frame for plotting}
\usage{
PrepareFragmentPlot(obj = NULL, counts = NULL, breaks, estimated.number = F)
}
\description{
prepare data frame for plotting
}
