% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{quickUmap}
\alias{quickUmap}
\title{run umap on PCA or other dm space}
\usage{
quickUmap(
  obj,
  space = "PCA",
  dims = NULL,
  n.neighbors = 30L,
  n.components = 2L,
  metric = "cosine",
  n.epochs = NULL,
  learning.rate = 1,
  min.dist = 0.3,
  spread = 1,
  set.op.mix.ratio = 1,
  local.connectivity = 1L,
  repulsion.strength = 1,
  negative.sample.rate = 5,
  a = NULL,
  b = NULL,
  uwot.sgd = FALSE,
  seed.use = 22,
  metric.kwds = NULL,
  angular.rp.forest = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{pagoda object}

\item{dims}{number of significant PCs for reduction (e.g. 1:20)}
}
\description{
run umap on PCA or other dm space
}
