% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read.bam.tags}
\alias{read.bam.tags}
\title{read bam files and extract tags from minimap2}
\usage{
read.bam.tags(filename, read.tag.names = F, fix.chromosome.names = F)
}
\arguments{
\item{filename}{bam file}
}
\description{
read bam files and extract tags from minimap2
}
