% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{RunJD}
\alias{RunJD}
\title{dimension reduction and visulization functions
caculate jaccard index matrix from an existing object}
\usage{
RunJD(obj, max.peaks = 1e+05, max.cell = 10000, seed.use = 10)
}
\arguments{
\item{max.peaks}{max number of peaks, above it will do sampling}

\item{max.cell}{max number of rows downsampled for calculating jaccard matrix}

\item{existing}{object: default pagoda2}
}
\description{
dimension reduction and visulization functions
caculate jaccard index matrix from an existing object
}
