% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{sampleBackgroundSeq}
\alias{sampleBackgroundSeq}
\title{sample background sequences based on mathed GC content distribution from peaks}
\usage{
sampleBackgroundSeq(
  obj,
  features,
  genome = "BSgenome.Hsapiens.NCBI.GRCh38",
  nsample = NULL
)
}
\arguments{
\item{features}{sets of features (peaks) for comparison (Grange object)}

\item{genome}{genome sequences}
}
\description{
sample background sequences based on mathed GC content distribution from peaks
}
