% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sampleCell}
\alias{sampleCell}
\title{sampling background cells}
\usage{
sampleCell(obj, posCell, method = c("knn", "random"), bgCellnumber = NULL)
}
\arguments{
\item{obj}{pagoda2 object}

\item{posCell}{position of positive cells}

\item{method, }{sample cells based on nearest cells or random cells}
}
\description{
sampling background cells
}
