% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet.R
\name{simulateDoubletMatrix}
\alias{simulateDoubletMatrix}
\title{simulate doublet matrix}
\usage{
simulateDoubletMatrix(
  mat,
  sampleRatio1 = 0.5,
  sampleRatio2 = 0.5,
  nSample = 1000,
  clusters = NULL,
  sampleCluster = TRUE,
  clusterRatio = 0.25
)
}
\description{
simulate doublet matrix
}
