% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{TSSenrichment}
\alias{TSSenrichment}
\title{Compute TSS enrichment score for each cell}
\usage{
TSSenrichment(
  obj,
  regions = NULL,
  flankbp = 100,
  sampleRegion = T,
  nsamples = 5000,
  n.cores = 20
)
}
\arguments{
\item{promoters}{GRanges object contains promoters regions (-1000 and +1000 flanking regions of a TSS)}
}
\description{
Compute TSS enrichment score for each cell
}
